requirements.txt
fastapi[all]
uvicorn
passlib
bcrypt==4.2.0
python-jose[cryptography]
psycopg2-binary
sqlalchemy